Move AccessList{,Member} validation to a separate file
